+++ src/main/java/ru/job4j/chat/controller/RoomController.java
@@ -3,6 +3,7 @@ package ru.job4j.chat.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.chat.domain.Room;
 import ru.job4j.chat.repository.RoomRepository;
 
@@ -29,15 +30,25 @@ public class RoomController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Room> findById(@PathVariable int id) {
+        if (id < 1) {
+            throw new NullPointerException("Room id can`t be less than 1");
+        }
         var room = this.roomRepository.findById(id);
         return new ResponseEntity<>(
-                room.orElse(new Room()),
-                room.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
+                room.orElseThrow(() ->
+                        new ResponseStatusException(
+                                HttpStatus.NOT_FOUND, "Room not found. Please, check id"
+                        )
+                ),
+                HttpStatus.OK
         );
     }
 
     @PostMapping("/")
     public ResponseEntity<Room> create(@RequestBody Room room) {
+        if (room.getName() == null || room.getDescription() == null) {
+            throw new NullPointerException("Room name and description can`t be empty");
+        }
         return new ResponseEntity<>(
                 this.roomRepository.save(room),
                 HttpStatus.CREATED
@@ -46,12 +57,18 @@ public class RoomController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Room room) {
+        if (room.getName() == null || room.getDescription() == null) {
+            throw new NullPointerException("Room name and description can`t be empty");
+        }
         this.roomRepository.save(room);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (id < 1) {
+            throw new NullPointerException("Room id can`t be less than 1");
+        }
         Room room = new Room();
         room.setId(id);
         this.roomRepository.delete(room);
